# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API для создания и управления вишлистами.",
        "title": "Wishlist API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/wishlists/{wishlistId}/gifts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gifts"
                ],
                "summary": "List wishlist gifts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wishlist ID",
                        "name": "wishlistId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gifts_handler.GiftsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gifts"
                ],
                "summary": "Create gift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wishlist ID",
                        "name": "wishlistId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Gift payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gifts_handler.CreateGiftInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_gifts_handler.CreateGiftOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/wishlists/{wishlistId}/gifts/{giftId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gifts"
                ],
                "summary": "Get gift by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wishlist ID",
                        "name": "wishlistId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gift ID",
                        "name": "giftId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gifts_handler.Gift"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gifts"
                ],
                "summary": "Update gift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wishlist ID",
                        "name": "wishlistId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gift ID",
                        "name": "giftId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Gift payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gifts_handler.UpdateGiftInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gifts"
                ],
                "summary": "Delete gift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wishlist ID",
                        "name": "wishlistId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gift ID",
                        "name": "giftId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorPayload": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_ChernykhITMO_Wishlist-API_internal_platform_httpcommon.ErrorBody"
                }
            }
        },
...
        "internal_gifts_handler.CreateGiftInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "internal_gifts_handler.CreateGiftOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "internal_gifts_handler.Gift": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "wishlistId": {
                    "type": "string"
                }
            }
        },
        "internal_gifts_handler.GiftsResponse": {
            "type": "object",
            "properties": {
                "gifts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_gifts_handler.Gift"
                    }
                }
            }
        },
        "internal_gifts_handler.UpdateGiftInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}